+++ CHANGELOG.md
@@ -30,6 +30,7 @@
 - Corporation Overview: No Data Error on DoesNotExist error
 - Add Char Button shown if memberaudit is active
 - Char/Corp Ledger Footer Amount changes wrong if sorting.
+- Billboard: Calculation Issue
 
 ## \[0.5.3\] - 2024-07-13